Исправление ошибок: разделение на классы
