+++ src/main/java/Controller/Controller.java
@@ -1,6 +1,7 @@
 package Controller;
 
-import Model.DataBase.DataBaseHandler;
+import Controller.ActionHandlers.ClientMainActionHandler;
+import Controller.ActionHandlers.ManagerMainActionHandler;
 import Model.Entities.Users.*;
 import Model.Exceptions.UserExc.*;
 import Model.LoggerUtil.Levels;
@@ -9,13 +10,14 @@ import Model.LoggerUtil.Logger;
 import Model.UserManagement.AuthenticationManager;
 import Model.UserManagement.Encryptor;
 import Model.UserManagement.RegistrationManager;
+import lombok.Getter;
 import ui.Menu;
-import ui.messageSrc.commands.*;
 import ui.out.Printer;
 import ui.messageSrc.Messages;
 
 /**
  * Главный класс, который отвечает за переключение и контроль сцен.
+ *
  * @see Scenes
  * Контроллер взаимодействует с классом Menu для отображения меню и подсказок пользователю.
  */
@@ -39,13 +41,10 @@ public class Controller extends Thread {
 
     private final ActionController actionController = new ActionController();
 
+    @Getter
     volatile User currentUser;
 
-    /**
-     * Переключатель задержки, она нужна для создания "анимации" и более привычного обращения с программой
-     */
-
-    boolean timeDelayOn = false;
+    public static Scenes currentScene = Scenes.currentScene;
 
     /**
      * Основной переключатель, который с помощью сцен переключается между действиями
@@ -53,112 +52,71 @@ public class Controller extends Thread {
      */
 
     public void run() {
-
         Controller.logger.log(Levels.INFO, "Система запущена!");
-        for(Scenes scene = Scenes.GREETING; checkSceneCycle(scene); ){
-            timeDelay(3000);
-            switch (scene) {
+        for (; ; ) {
+            switch (currentScene) {
 
-                case GREETING -> scene = greeting();
+                case GREETING -> greeting();
 
-                case CHOOSING_ROLE -> scene = chooseRegistrationOrAuth();
+                case CHOOSING_ROLE -> chooseRegistrationOrAuth();
 
-                case REGISTRATION -> scene = registrationHandler();
+                case ACTIONS -> actionController.distributeActionsByRoots();
 
-                case AUTHORIZATION -> scene = authorizationHandler();
+                case EXIT_FROM_ACCOUNT -> logOut();
 
-                case ACTIONS -> scene = actionController.distributeActions();
-
-                case EXIT_FROM_ACCOUNT -> scene = logOut();
-
-                case SHUT_DOWN -> scene = Scenes.ACTIONS;
+                case SHUT_DOWN -> {
+                    return;
+                }
             }
         }
     }
 
-    private boolean checkSceneCycle(Scenes scene){
-        boolean isShutDown = !(scene.equals(Scenes.SHUT_DOWN) && Menu.areYouSure(Messages.SHUT_DOWN_WARNING.getMessage()));
-        if(!isShutDown) Controller.logger.log(Levels.INFO, "Система отключается администратором!");
-        return isShutDown;
-    }
-
-    public Scenes greeting(){
+    public void greeting() {
         Menu.greeting();
-        return Scenes.GREETING.nextStep();
+        currentScene.nextStep();
     }
 
-    public Scenes chooseRegistrationOrAuth(){
+    public void chooseRegistrationOrAuth() {
         Printer.print(Messages.ACTIONS_TO_ENTER.getMessage());
-        return (Menu.chooseRegistrationOrAuth().equals("Войти")) ? Scenes.AUTHORIZATION : Scenes.REGISTRATION;
+        signHandler();
     }
 
-    public Scenes registrationHandler(){
+    public void signHandler() {
         try {
-            currentUser = registration();
-            timeDelay(200);
-            Printer.print("Вы создали аккаунт с ID " + currentUser.getID() + " и именем " + currentUser.getName() + " Ваша роль: " + currentUser.getAccessLevel().getValue());
+            currentUser = (Menu.chooseRegistrationOrAuth().equals("Sing in")) ? authorization() : registration();
+            Printer.print("Successfully! Your ID " + currentUser.getID() + "\t Name: " + currentUser.getName() + "\t Role: " + currentUser.getAccessLevel().getValue());
             logger.log(Levels.INFO, LogActions.USER_REGISTERED.getText() + currentUser.toString());
-            return Scenes.ACTIONS;
-        } catch (RegistrationInterruptException e) {
-            Printer.print(Messages.ERROR.getMessage());
-        } catch (UserAlreadyExistsException e) {
-            Printer.print(Messages.USER_ALREADY_EXISTS.getMessage());
-        }
-        return Scenes.CHOOSING_ROLE;
-    }
-
-    public Scenes authorizationHandler(){
-        try {
-            authorization();
-            timeDelay(2000);
-            Printer.print("Вы вошли в аккаунт под ID " + currentUser.getID() + " и именем " + currentUser.getName() + " Ваша роль: " + currentUser.getAccessLevel().getValue());
-            logger.log(Levels.INFO, LogActions.USER_AUTHORIZED.getText() + currentUser.toString());
-            return Scenes.ACTIONS;
-        } catch (InvalidPasswordException e){
-            Printer.print(Messages.INVALID_PASS.getMessage());
-        } catch (NoSuchUserException e) {
-            Printer.print(Messages.NO_SUCH_USER.getMessage());
+            currentScene = Scenes.ACTIONS;
+            return;
         } catch (Exception e) {
-            Printer.print(Messages.ERROR.getMessage());
+            Printer.print("Invalid input data, try again!");
         }
-        return Scenes.CHOOSING_ROLE;
+        currentScene = Scenes.CHOOSING_ROLE;
     }
 
     private User registration() throws RegistrationInterruptException, UserAlreadyExistsException {
-        User newUser = RegistrationManager.registration(
+        User user = RegistrationManager.registration(
                 Menu.chooseRole(),
                 Menu.getUserName(),
                 Menu.getUserPassword()
         );
-        if(newUser instanceof Client || newUser instanceof Manager)
-            newUser.setPhoneNumber(Menu.getUserPhoneNumber());
-        return newUser;
+        if (user instanceof Client || user instanceof Manager)
+            user.setPhoneNumber(Menu.getUserPhoneNumber());
+        return user;
     }
 
-    private void authorization() throws Exception, InvalidPasswordException {
-        Printer.printCentered("Добро пожаловать!");
-        currentUser = AuthenticationManager.authentication(
+    private User authorization() throws Exception {
+        return AuthenticationManager.authentication(
                 Menu.getUserName(),
                 Encryptor.encrypt(Menu.getUserPassword())
         );
     }
 
-    public Scenes logOut(){
+    public void logOut() {
         logger.log(Levels.INFO, LogActions.USER_EXIT.getText() + currentUser.toString());
         currentUser = null;
-        return Scenes.CHOOSING_ROLE;
+        currentScene = Scenes.CHOOSING_ROLE;
     }
-
-    private void timeDelay(int ms){
-        if(!timeDelayOn) return;
-        try{
-            Thread.sleep(ms);
-        } catch (InterruptedException ignored){
-            logger.log(Levels.ERR, LogActions.ERROR.getText() + " прерывание искусственной задержки");
-            System.err.println("Задержка прервана");
-        }
-    }
-
 
     /**
      * Основной контроллер, который переключает действия в зависимости от уровня доступа
@@ -166,63 +124,14 @@ public class Controller extends Thread {
 
     private class ActionController {
 
-        private Scenes distributeActions(){
-            Scenes nextScene = Scenes.ACTIONS;
-            try {
-                switch (currentUser.getAccessLevel()) {
-                    case CLIENT -> nextScene = clientActionHandler(Menu.clientChoosingAction());
-                    case MANAGER -> nextScene = managerActionHandler(Menu.managerChoosingAction());
-                    case ADMINISTRATOR -> nextScene = adminActionHandler(Menu.adminChoosingAction());
-                }
-            } catch (NotEnoughRightsException rightsException){
-                Printer.print("Неверное действие, проверьте корректность введенных данных и попробуйте снова");
-                distributeActions();
+        private void distributeActionsByRoots() {
+            currentScene = Scenes.ACTIONS;
+            switch (currentUser.getAccessLevel()) {
+                case CLIENT -> ClientMainActionHandler.distribute(Menu.clientChoosingAction());
+                case MANAGER -> ManagerMainActionHandler.distribute(Menu.managerChoosingAction());
+//                case ADMINISTRATOR ->
             }
-            return nextScene;
         }
 
-        private Scenes clientActionHandler(ClientCommands action) throws NotEnoughRightsException {
-            if(!currentUser.getAccessLevel().equals(AccessLevels.CLIENT)) throw new NotEnoughRightsException();
-            Scenes nextScene = Scenes.ACTIONS;
-            switch (action) {
-                case VIEW_ORDERS -> ActionHandler.viewUserOrders((Client) currentUser);
-                case VIEW_USER_CARS -> ActionHandler.viewUserCars((Client) currentUser);
-                case ADD_USER_CAR -> ActionHandler.addUserCar((Client) currentUser);
-                case REMOVE_USER_CAR -> ActionHandler.removeUserCar((Client) currentUser);
-                case GO_TO_SHOWROOM -> ActionHandler.goToShowRoom(currentUser);
-                case SETUP_MY_PROFILE -> nextScene = ActionHandler.setUpUserParameters(currentUser.getID());
-                case EXIT_FROM_ACCOUNT -> nextScene = Scenes.EXIT_FROM_ACCOUNT;
-                case DELETE_ACCOUNT -> nextScene = ActionHandler.removeAccount(currentUser);
-                default -> throw new NotEnoughRightsException();
-            }
-            return nextScene;
-        }
-
-        private Scenes managerActionHandler(ManagerCommands action) throws NotEnoughRightsException {
-            Scenes nextScene = Scenes.ACTIONS;
-            switch (action) {
-                case ORDERS -> ActionHandler.gotoOrdersPage((Manager) currentUser);
-                case EXIT_FROM_ACCOUNT -> nextScene = Scenes.EXIT_FROM_ACCOUNT;
-                case GO_TO_SHOWROOM -> ActionHandler.goToShowRoom(currentUser);
-                case SETUP_MY_PROFILE -> nextScene = ActionHandler.setUpUserParameters(currentUser.getID());
-                case DELETE_ACCOUNT -> nextScene = ActionHandler.removeAccount(currentUser);
-            }
-            return nextScene;
-        }
-
-        private Scenes adminActionHandler(AdminCommands action) throws NotEnoughRightsException {
-            Scenes nextScene = Scenes.ACTIONS;
-            switch (action) {
-                case GO_TO_USER_LIST ->  ActionHandler.gotoUserListPage((Administrator) currentUser);
-                case GET_LOG_LIST -> ActionHandler.getLogList();
-                case SAVE_LOG_LIST -> ActionHandler.saveLogList();
-                case DELETE_ACCOUNT -> nextScene = ActionHandler.removeAccount(currentUser);
-                case SETUP_MY_PROFILE -> nextScene = ActionHandler.setUpUserParameters(currentUser.getID());
-                case EXIT_FROM_ACCOUNT -> nextScene = Scenes.EXIT_FROM_ACCOUNT;
-                case SHUT_DOWN -> nextScene = Scenes.SHUT_DOWN;
-            }
-            return nextScene;
-        }
     }
-
 }
