+++ src/main/java/Controller/Scenes.java
@@ -19,13 +19,21 @@ public enum Scenes {
 
     @Getter
     private int number;
-
+    @Getter
+    public static Scenes currentScene = GREETING;
     Scenes(int number){
         this.number = number;
     }
 
-    Scenes nextStep(){
-        if(number==8) return Scenes.SHUT_DOWN;
-        return Scenes.values()[number];
+    /**
+     * Переключение сцены на один шаг вперед.
+     * Так как нумерация списка начинается с 0, а нумерация сцен с 1,
+     * то единицу прибавлять не нужно для следующего шага,
+     * достаточно взять элемент из списка по значению сцены
+     */
+
+    void nextStep(){
+        if(!currentScene.equals(SHUT_DOWN))
+            currentScene = Scenes.values()[currentScene.ordinal()];
     }
 }
