+++ src/main/java/Model/Exceptions/UserExc/InvalidPasswordException.java
@@ -4,7 +4,7 @@ import Controller.Controller;
 import Model.LoggerUtil.Levels;
 import Model.LoggerUtil.LogActions;
 
-public class InvalidPasswordException extends Throwable {
+public class InvalidPasswordException extends Exception {
     public InvalidPasswordException() {
         Controller.logger.log(Levels.ERR, LogActions.ERROR.getText() + " Неверный пароль!");
     }
